+++ src/eulertools/subcommands/run.py
@@ -8,6 +8,7 @@ from eulertools.lib.exceptions import FailedRunError
 from eulertools.lib.utils import (
     Language,
     Modes,
+    Problem,
     get_answers,
     get_line_answer,
     get_line_timing,
@@ -20,7 +21,7 @@ class Run:
     def __init__(
         self,
         languages: list[Language],
-        problems: list[str],
+        problems: list[Problem],
         verbosity: int,
         *,
         run_update: bool = False,
@@ -39,19 +40,19 @@ class Run:
         output: dict[Language, dict[str, dict[int, list[Timing]]]] = {}
         for language, problem in product(self.languages, self.problems):
             timings = self.run_single_problem(language, problem)
-            output.setdefault(language, {})[problem] = timings
+            output.setdefault(language, {})[problem.id] = timings
         if self.run_update:
             update_answers(self.expected_answers)
         return output
 
     def run_single_problem(
-        self, language: Language, problem: str
+        self, language: Language, problem: Problem
     ) -> dict[int, list[Timing]]:
         solution = get_solution(language, problem)
         if not solution.exists():
             return {}
         raw_output = subprocess.run(
-            [language.runner, problem, str(self.times)],  # noqa: S603
+            [language.runner, problem.id, str(self.times)],  # noqa: S603
             capture_output=True,
             check=True,
         )
@@ -69,20 +70,20 @@ class Run:
                 actual_answers.setdefault(run_id, set()).add(answer)
             else:
                 print(
-                    f"🔴 Running {language.name}/{problem}... Cannot parse `{line}`.",
+                    f"🔴 Running {language.name} // {problem.id}... Cannot parse `{line}`.",
                     file=sys.stderr,
                 )
-                msg = f"Unsuccessful run of {language.name}/{problem}"
+                msg = f"Unsuccessful run of {language.name} // {problem.id}"
                 raise FailedRunError(msg)
-        expected_answers = self.expected_answers.setdefault(problem, {})
+        expected_answers = self.expected_answers.setdefault(problem.id, {})
         if missing_answers := {
             answer for answer in expected_answers if answer not in actual_answers
         }:
             print(
-                f"🔴 Running {language.name}/{problem}... Missing answers with keys {missing_answers}.",
+                f"🔴 Running {language.name} // {problem.id}... Missing answers with keys {missing_answers}.",
                 file=sys.stderr,
             )
-            msg = f"Unsuccessful run of {language.name}/{problem}"
+            msg = f"Unsuccessful run of {language.name} // {problem.id}"
             raise FailedRunError(msg)
         success = True
         for key, values in actual_answers.items():
@@ -90,24 +91,24 @@ class Run:
             if len(values) != 0:
                 success = False
                 print(
-                    f"🔴 Running {language.name}/{problem}/{key}... Not deterministic answer.",
+                    f"🔴 Running {language.name} // {problem.id} // {key}... Not deterministic answer.",
                     file=sys.stderr,
                 )
             elif key not in expected_answers:
                 if self.mode != Modes.TIMING:
                     print(
-                        f"🟠 Running {language.name}/{problem}/{key}... new response: {value}"
+                        f"🟠 Running {language.name} // {problem.id} // {key}... new response: {value}"
                     )
                 expected_answers[key] = value
             elif value != expected_answers[key]:
                 success = False
                 print(
-                    f"🔴 Running {language.name}/{problem}/{key}... expected: {expected_answers[key]}, got: {value}",
+                    f"🔴 Running {language.name} // {problem.id} // {key}... expected: {expected_answers[key]}, got: {value}",
                     file=sys.stderr,
                 )
             elif self.mode != Modes.TIMING:
-                print(f"🟢 Running {language.name}/{problem}/{key}... {value}")
+                print(f"🟢 Running {language.name} // {problem.id} // {key}... {value}")
         if not success:
-            msg = f"Unsuccessful run of {language.name}/{problem}"
+            msg = f"Unsuccessful run of {language.name} // {problem.id}"
             raise FailedRunError(msg)
         return timings