+++ src/eulertools/subcommands/test.py
@@ -1,12 +1,16 @@
 from itertools import product
 
-from eulertools.lib.utils import Language, Modes, get_solution
+from eulertools.lib.utils import Language, Modes, Problem, get_solution
 from eulertools.subcommands.run import Run
 
 
 class Test:
     def __init__(
-        self, languages: list[Language], problems: list[str], times: int, verbosity: int
+        self,
+        languages: list[Language],
+        problems: list[Problem],
+        times: int,
+        verbosity: int,
     ):
         self.languages = languages
         self.problems = problems
@@ -17,7 +21,7 @@ class Test:
         for language, problem in product(self.languages, self.problems):
             self.test_single_problem(language, problem)
 
-    def test_single_problem(self, language: Language, problem: str) -> None:
+    def test_single_problem(self, language: Language, problem: Problem) -> None:
         solution = get_solution(language, problem)
         if not solution.exists():
             return
@@ -28,4 +32,4 @@ class Test:
             verbosity=self.verbosity,
             mode=Modes.RUN,
             times=self.times,
-        ).run()[language][problem]
+        ).run()[language][problem.id]