+++ docs/CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ### Added
 
--   Added a way to allow passing the problem name in a different format
+-   Added a way to allow passing the problem name with an id
 
 ## [2.1.0] - 2023-12-06